Extract PDF download related code to module
