+++ src/ringgo/mod.rs
@@ -1,4 +1,5 @@
 pub mod authentication;
+pub mod downloads;
 pub mod errors;
 pub mod parking_sessions;
 pub mod url_helpers;
